gamescene skeleton
The neon tetra (Paracheirodon innesi) is a Freshwater fish of the characin family (family Characidae) of order Characiformes. The type species of its genus, it is native to blackwater and clearwater streams in the Amazon basin of South America. Its bright colouring makes the fish visible to conspecifics in the dark blackwater streams, and is also the main reason for its popularity among freshwater fish hobbyists, with neon tetras being one of the most widely kept tropical fish in the world.